feat: Web css & directive template
user's input strike count, ball count form aligned.
And angular template in directive whether it is the guess data of the bot or the user's input data, and changes the template accordingly. it declared a function that could work in the template through the link in directive and handled it with the processing of the template through ng-hide tag attribute.