feat: Add the 3-strkike, 3-ball and focus input text
change the range from 0 to 3 for existing strikies and ball counts from 0 to 2. if it is a 3 strkie, the result is displayed only by changing the game status to the 'over' state. Also, they are preparing a new game case.
the autofocus on the input form uses the  function of the form in the directive of the angular. the  function is performed by changing the value of the bool every time input is completed.